+++ src/components/Header.vue
@@ -1,7 +1,8 @@
 <template>
-    <div class="headerComponent">
-        <img src="@/assets/DSWDLogo.png" alt="dswdLogo">
-    </div>
+  
+    <header class="d-flex flex-wrap justify-content-start  ">
+      <img src="../assets/DSWDLogo.png" alt="">
+    </header>
   </template>
   
   <script>
@@ -11,12 +12,14 @@
   </script>
   
   <style scoped>
-
+    /* .header-image{
+      padding-left: 2em;
+    }
     .headerComponent{
       position: sticky;
       
       height: 80px;
-      background: linear-gradient(to right, #EE1C25, white, #294D9C);
+      background: linear-gradient(to right, white 15%, #EE1C25 85%);
 
       z-index: 999;
       margin-top:-60px;
@@ -27,8 +30,23 @@
     .headerComponent img{
       height: 70px;
       padding-top: 10px;
+    } */
+
+
+    header{
+      background: linear-gradient(to right, white 20%, #EE1C25 80%);
+      width: 100%;
+      position: fixed;
+      top: 0;
+  
     }
 
+    header img{
+      width: 280px;
+   
+      margin: 20px 0px 10px 20px;
+      
+    }
 
   </style>