+++ src/views/LoginForm.vue
@@ -1,6 +1,12 @@
 <template>
-<Header />
+<!-- <Header /> -->
 <div class="parent clearfix">
+
+    <div class="headerComponent">
+        <img class="float-start header-image" src="@/assets/DSWDLogo.png" alt="dswdLogo">
+    </div>
+
+
     <div class="bg-illustration">
 
         <!-- <img src="https://i.ibb.co/Pcg0Pk1/logo.png" alt="logo"> -->
@@ -72,12 +78,11 @@
 
 <script>
 import Footer from '@/components/Footer'; 
-import Header from '@/components/Header'
 
 
 export default {
   name: 'LoginForm',
-  components: { Footer, Header },
+  components: { Footer},
   data() {
   return {
     username: '',
@@ -91,24 +96,32 @@ export default {
 
 methods: {
   login() {
-    // Set loading state to true
-    this.loading = true;
+  // Set loading state to true
+  this.loading = true;
 
-    // Simulate a delay to mimic a server response (remove this in a real scenario)
-    setTimeout(() => {
-      // Perform authentication logic here (e.g., check username and password)
-      if (this.username === 'Admin' && this.password === 'admin') {
-        // Authentication successful, navigate to the dashboard
-        this.$router.push('/dashboard');
-      } else {
-        // Authentication failed, display error message
-        this.error = true;
-      }
+  // Simulate a delay to mimic a server response (remove this in a real scenario)
+  setTimeout(() => {
+    // Perform authentication logic here (e.g., check username and password)
+    if (this.username === 'Admin' && this.password === 'admin') {
+      // Authentication successful, set user as authenticated in localStorage
+      localStorage.setItem('user', 'authenticated');
 
-      // Reset loading state to false after authentication logic
-      this.loading = false;
-    }, 1000); // Simulated delay of 1 second
+      // Navigate to the dashboard
+      this.$router.push('/dashboard');
+    } else {
+      // Authentication failed, display error message
+      this.error = true;
+    }
+
+    // Reset loading state to false after authentication logic
+    this.loading = false;
+  }, 1000); // Simulated delay of 1 second
+},
+created() {
+    // Clear the user's authentication status in localStorage when the login page is visited
+    localStorage.removeItem('user');
   },
+
 },
 
 };
@@ -132,6 +145,7 @@ methods: {
   /* padding-top: 10px;  */
   padding-bottom: 10px;
   width: 100%; /* Set a default width */
+  visibility: hidden;
 
   @media only screen and (min-width: 400px) {
     width: 60%;
@@ -156,16 +170,16 @@ methods: {
   }
 
   @media only screen and (min-width: 1024px) {
+    visibility: visible;
     padding-top: 40px;
     margin-left: 20px; /* Adjust margin if necessary */
     width: 90%;
   }
+
+  @media only screen and (min-width: 1244px) {
+
+  }
 }
-
-
-
-
-
 
 
 
@@ -179,6 +193,29 @@ methods: {
   -o-transition: 0.3s;
   transition: 0.3s;
 }
+
+
+    .header-image{
+      padding-left: 50px;
+    }
+    .headerComponent{
+      position: relative;
+
+      height: 5em;
+      background: linear-gradient(to right, white 15%, #EE1C25 85%);
+      z-index: 999;
+      margin-top:-60px;
+      margin-bottom: 20px;
+      padding: 0;
+    }
+
+    .headerComponent img{
+      height: 70px;
+      padding-top: 10px;
+    }
+
+
+
 
 body {
   background-color: #fff;
@@ -238,7 +275,7 @@ img {
 .bg-illustration {
   margin-top: -3.73em;
   position: relative;
-  height: 91vh;
+  height: 100vh;
   width: 1194px;
   /* background: url("https://i.ibb.co/RhMZprS/jony-Image2-1.png") no-repeat center center scroll; */
   background: #294D9C;
@@ -246,6 +283,15 @@ img {
   float: left;
   -webkit-animation: bgslide 2.3s forwards;
           animation: bgslide 2.3s forwards;
+
+
+ @media only screen and (min-width: 1280px) {
+    height: 95vh; 
+  }
+
+  @media only screen and (min-width: 2160px) {
+    height: 100vh; 
+  }
 }
 .bg-illustration img {
   width: 248px;
@@ -293,11 +339,16 @@ img {
 
 
 .login {
-  max-height: 90vh;
+  max-height: 70vh;
   overflow-Y: auto;
   float: left;
   margin: 0 auto;
   width: calc(100% - 1194px);
+
+@media only screen and (min-width: 1024px) {
+  max-height: 100vh;
+}
+
 }
 .login .container {
   width: 505px;