+++ src/views/SLP.vue
@@ -1,0 +1,23 @@
+<template>
+    <Sidebar />
+
+    <br><br><br><br>
+
+    <h1>SLP DASHBOARD</h1>
+
+    <!-- <Footer /> -->
+</template>
+
+<script>
+import Sidebar from '@/components/Sidebar.vue'; 
+import Footer from '@/components/Footer';
+
+export default {
+    name: 'SLP',
+    components: {
+        Sidebar,
+        Footer,
+    }
+    
+}
+</script>