+++ CHANGELOG.md
@@ -14,4 +14,16 @@
 
 ## ver 0.6 edited login page, added footer, header, image carousel and text template - MARLO BARUA
 
-## ver 1.2 Added Employment data on OSP - MARA SUERTE
+## ver 0.7 optimize the mobile version of the login page and the dashboard header - MARLO BARUA
+
+## ver 0.8 Added multiple charts (Polar, Doughnut, Area), Added SLP Dashboard - MARLO BARUA
+
+## ver 0.9 Customize the BarChart so that It will read multiple data with props - MARLO BARUA
+
+## ver 0.10 Customize (Line, Doughnut, Pie Charts) so that it will read multiple data properties with vue props - MARLO BARUA
+
+## ver 1.0 Partially done with the header and sidebar for the dashboard, temporarily removed the footer for some fixings, Fix the route guard protection for authentication - MARLO BARUA
+
+## ver 1.1 Added Horizontal Bar Chart, Started the Development for SWDA Dashboard - MARLO BARUA
+
+## ver 1.1.1 Fix the footer in Login Page and Input details in SWDA Dashboard - MARLO BARUA